take LiveAtom.full_density to be view2D<double>

+++ src/inline_math.hxx
@@ -39,60 +39,60 @@
 
 
   template<typename real_t>
-  void inline set(real_t y[], int const n, real_t const a) {
-      for(int i = 0; i < n; ++i) y[i] = a;
+  void inline set(real_t y[], size_t const n, real_t const a) {
+      for(size_t i = 0; i < n; ++i) y[i] = a;
   } // set
 
   template<typename real_t>
-  void inline set(real_t y[], int const n, real_t const a[], real_t const f=1) {
-      for(int i = 0; i < n; ++i) y[i] = a[i]*f;
+  void inline set(real_t y[], size_t const n, real_t const a[], real_t const f=1) {
+      for(size_t i = 0; i < n; ++i) y[i] = a[i]*f;
   } // set
 
   template<typename real_t>
-  void inline scale(real_t y[], int const n, real_t const a[], real_t const f=1) {
-      for(int i = 0; i < n; ++i) y[i] *= a[i]*f;
+  void inline scale(real_t y[], size_t const n, real_t const a[], real_t const f=1) {
+      for(size_t i = 0; i < n; ++i) y[i] *= a[i]*f;
   } // scale
 
   template<typename real_t>
-  void inline scale(real_t y[], int const n, real_t const f) {
+  void inline scale(real_t y[], size_t const n, real_t const f) {
       // no default value for f here, as scaling with 1.0 has no effect
-      for(int i = 0; i < n; ++i) y[i] *= f;
+      for(size_t i = 0; i < n; ++i) y[i] *= f;
   } // scale
 
   template<typename real_t>
-  void inline product(real_t y[], int const n, real_t const a[], real_t const b[], real_t const f=1) {
-      for(int i = 0; i < n; ++i) y[i] = a[i]*b[i]*f;
+  void inline product(real_t y[], size_t const n, real_t const a[], real_t const b[], real_t const f=1) {
+      for(size_t i = 0; i < n; ++i) y[i] = a[i]*b[i]*f;
   } // product
 
   template<typename real_t>
-  void inline product(real_t y[], int const n, real_t const a[], real_t const b[], real_t const c[], real_t const f=1) {
-      for(int i = 0; i < n; ++i) y[i] = a[i]*b[i]*c[i]*f;
+  void inline product(real_t y[], size_t const n, real_t const a[], real_t const b[], real_t const c[], real_t const f=1) {
+      for(size_t i = 0; i < n; ++i) y[i] = a[i]*b[i]*c[i]*f;
   } // product
 
   template<typename real_t>
-  void inline add_product(real_t y[], int const n, real_t const a[], real_t const f) {
+  void inline add_product(real_t y[], size_t const n, real_t const a[], real_t const f) {
       // no default value for f here, otherwise the name add_product is missleading!
-      for(int i = 0; i < n; ++i) y[i] += a[i]*f;
+      for(size_t i = 0; i < n; ++i) y[i] += a[i]*f;
   } // add_product == axpy-type
 
   template<typename real_t>
-  void inline add_product(real_t y[], int const n, real_t const a[], real_t const b[], real_t const f=1) {
-      for(int i = 0; i < n; ++i) y[i] += a[i]*b[i]*f;
+  void inline add_product(real_t y[], size_t const n, real_t const a[], real_t const b[], real_t const f=1) {
+      for(size_t i = 0; i < n; ++i) y[i] += a[i]*b[i]*f;
   } // add_product
 
   template<typename real_t>
-  real_t inline dot_product(int const n, real_t const bra[], real_t const ket[]) {
+  real_t inline dot_product(size_t const n, real_t const bra[], real_t const ket[]) {
       real_t dot = 0;
-      for(int i = 0; i < n; ++i) {
+      for(size_t i = 0; i < n; ++i) {
           dot += bra[i]*ket[i];
       } // i
       return dot;
   } // dot_product
 
   template<typename real_t>
-  real_t inline dot_product(int const n, real_t const bra[], real_t const ket[], real_t const metric[]) {
+  real_t inline dot_product(size_t const n, real_t const bra[], real_t const ket[], real_t const metric[]) {
       real_t dot = 0;
-      for(int i = 0; i < n; ++i) {
+      for(size_t i = 0; i < n; ++i) {
           dot += bra[i]*metric[i]*ket[i];
       } // i
       return dot;
